Revert "Pydantic try 2"
