+++ .github/workflows/main.yml
@@ -22,4 +22,4 @@ jobs:
         run: ls -la
 
       - name: Run tests
-        run: pytest -v
+        run: pytest testes.py -v
