+++ pyproject.toml
@@ -19,6 +19,6 @@ build-backend = "poetry.core.masonry.api"
 [tool.taskipy.tasks]
 
 
-run = "uvicorn app.main:app --reload"
+run = "uvicorn main:app --reload"
 test = "pytest -v"
 kill = "kill -9 $(lsof -t -i:8000)"
